requirements.txt

python-telegram-bot>=20.7
aiohttp>=3.9.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
cryptography>=41.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
async-timeout>=4.0.0
aiosqlite>=0.18.0